fix: update cast BaseDocumentRepostory

+++ jnosql-mapping/jnosql-mapping-column/src/main/java/org/eclipse/jnosql/mapping/column/query/BaseColumnRepository.java
@@ -158,7 +158,7 @@ public abstract class BaseColumnRepository<T> {
             sorts.addAll(query.sorts());
             sorts.addAll(special.sorts());
             long skip = limit.map(l -> l.startAt() - 1).orElse(query.skip());
-            long max = limit.map(Limit::maxResults).orElse(query.limit());
+            long max = limit.map(Limit::maxResults).orElse((int) query.limit());
             return new MappingColumnQuery(sorts, max,
                     skip,
                     query.condition().orElse(null),
@@ -167,7 +167,7 @@ public abstract class BaseColumnRepository<T> {
 
         if (limit.isPresent()) {
             long skip = limit.map(l -> l.startAt() - 1).orElse(query.skip());
-            long max = limit.map(Limit::maxResults).orElse(query.limit());
+            long max = limit.map(Limit::maxResults).orElse((int) query.limit());
             return new MappingColumnQuery(query.sorts(), max,
                     skip,
                     query.condition().orElse(null),
